+++ src/protocolsupport/protocol/packet/middleimpl/clientbound/play/v_4_5_6/SpawnObject.java
@@ -26,7 +26,7 @@ public class SpawnObject extends MiddleSpawnObject {
 				switch (objectdata) {
 					case 0: {
 						z -= 32;
-						yaw = (byte) 128;
+						yaw = 128;
 						break;
 					}
 					case 1: {
@@ -41,7 +41,7 @@ public class SpawnObject extends MiddleSpawnObject {
 					}
 					case 3: {
 						x -= 32;
-						yaw = (byte) 192;
+						yaw = 192;
 						break;
 					}
 				}
